Resolve effective address in opcode dispatch
resolve_address was taking ~10% of runtime to match against an
addressing mode that we had pinned down to a single variant earlier in
the program execution.

This patch makes it so the instruction's effective address is resolved
in the opcode dispatch switch where we know exactly what addressing
mode is being used, effectively removing the work of the match
statement.

This improves performance by ~10% on the Klaus Dormann functional test
program.